Move pagination methods to list VM
